+++ main.go
@@ -12,12 +12,15 @@ import (
 )
 
 func main() {
-	godotenv.Load()
+	err := godotenv.Load()
+	if err != nil {
+		logrus.Infof("failed to load environment variables from .env: %s", err.Error())
+	}
 	logrus.SetLevel(logrus.InfoLevel)
 	logrus.SetFormatter(&log.Formatter{})
 
 	startTime := time.Now()
-	if err := cmd.Execute(); err != nil {
+	if err = cmd.Execute(); err != nil {
 		if cmd.Interactive {
 			_ = beeep.Alert("Happy Error!", err.Error(), "assets/warning.png")
 		}