+++ pkg/backend/aws/ecs_compute_backend.go
@@ -1,0 +1,40 @@
+package aws
+
+import (
+	"context"
+	"encoding/json"
+
+	"github.com/aws/aws-sdk-go-v2/service/secretsmanager"
+	"github.com/chanzuckerberg/happy/pkg/backend/aws/interfaces"
+	"github.com/chanzuckerberg/happy/pkg/config"
+	"github.com/pkg/errors"
+)
+
+type ECSComputeBackend struct {
+	Backend     *Backend
+	HappyConfig *config.HappyConfig
+}
+
+func NewECSComputeBackend(ctx context.Context, happyConfig *config.HappyConfig, b *Backend) (interfaces.ComputeBackend, error) {
+	return &ECSComputeBackend{
+		Backend:     b,
+		HappyConfig: happyConfig,
+	}, nil
+}
+
+func (b *ECSComputeBackend) GetIntegrationSecret(ctx context.Context) (*config.IntegrationSecret, *string, error) {
+	secretId := b.HappyConfig.GetSecretId()
+	out, err := b.Backend.secretsclient.GetSecretValue(ctx, &secretsmanager.GetSecretValueInput{
+		SecretId: &secretId,
+	})
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "could not get integration secret at %s", secretId)
+	}
+
+	secret := &config.IntegrationSecret{}
+	err = json.Unmarshal([]byte(*out.SecretString), secret)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "could not json parse integraiton secret")
+	}
+	return secret, out.ARN, nil
+}